requirements-new.txt
# M-Bus MQTT Gateway - Production Dependencies

# Core Dependencies
paho-mqtt>=2.0.0                 # MQTT Client with async support
python-meterbus>=0.7.0           # M-Bus Protocol Library
pyserial>=3.5                    # Serial Port Communication

# Async & Performance
aiofiles>=23.0.0                 # Async File I/O
asyncio-mqtt>=0.16.0             # Async MQTT wrapper

# Data & Persistence
aiosqlite>=0.19.0                # Async SQLite
pydantic>=2.0.0                  # Data Validation & Settings

# Monitoring & Observability
aiohttp>=3.9.0                   # Async HTTP Server for Health Checks
prometheus-client>=0.19.0        # Metrics Export
structlog>=23.0.0                # Structured Logging

# Utilities
python-dotenv>=1.0.0             # Environment Variables
PyYAML>=6.0                      # YAML Config Support
tenacity>=8.2.0                  # Retry Logic with Backoff

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0
